refactor: 미사용 임포트 제거, 구조 개선, 파일명 통일화
