+++ src/main/java/com/goorm/team9/icontact/domain/block/dto/request/BlockRequestDto.java
@@ -1,4 +1,4 @@
-package com.goorm.team9.icontact.domain.block.dto;
+package com.goorm.team9.icontact.domain.block.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
@@ -13,4 +13,5 @@ public class BlockRequestDto {
 
     @Schema(example = "Noah2", description = "차단될 사용자 닉네임")
     private String blockedNickname;
+
 }
