+++ src/main/java/com/goorm/team9/icontact/domain/chat/dto/request/ChatRequestNotificationDto.java
@@ -1,4 +1,4 @@
-package com.goorm.team9.icontact.domain.chat.dto;
+package com.goorm.team9.icontact.domain.chat.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -8,7 +8,9 @@ import java.time.LocalDateTime;
 @Getter
 @AllArgsConstructor
 public class ChatRequestNotificationDto {
+
     private String senderNickname;
     private Long requestId;
     private LocalDateTime requestedAt;
+
 }
