+++ src/main/java/com/goorm/team9/icontact/domain/chat/dto/response/ChatMessageNotificationDto.java
@@ -1,4 +1,4 @@
-package com.goorm.team9.icontact.domain.chat.dto;
+package com.goorm.team9.icontact.domain.chat.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -8,8 +8,10 @@ import java.time.LocalDateTime;
 @Getter
 @AllArgsConstructor
 public class ChatMessageNotificationDto {
+
     private String senderNickname;
     private String messagePreview;
     private Long roomId;
     private LocalDateTime sentAt;
+
 }
